Monte Carlo - Callable Futures & no more BigDecimal

+++ src/main/java/com/adrian/ng/RiskMeasure.java
@@ -3,21 +3,20 @@ package com.adrian.ng;
 import yahoofinance.Stock;
 
 import java.io.IOException;
-import java.math.BigDecimal;
 
 public abstract class RiskMeasure extends VaR {
 
-    public int getSize(){
+    public int getSize() {
         Stock stck = stockHashMap.get(strSymbols[0]);
         int size = 0;
-        try{
-            size = stck.getHistory().size() -1;
-        }catch(IOException e){
+        try {
+            size = stck.getHistory().size() - 1;
+        } catch (IOException e) {
             e.printStackTrace();
         }
         return size;
     }
 
-    abstract BigDecimal getVar();
+    abstract double getVar();
 
 }
